Move acrobot monte carlo simulation out of dev
 * Toward #13904
 * This is the infrastructure to run monte carlo simulations
   of the acrobot under Spong control.
 * This is an example of how to use the stochastic schema
   mechanism for monte carlo testing
 * A subsequent PR will bring in an example of using this along
   with an optimizer to learn controller parameters (that code
   remains in dev in this PR)
