+++ examples/acrobot/dev/BUILD.bazel
@@ -17,14 +17,6 @@ load("//tools/lint:lint.bzl", "add_lint_tests")
 # Now we have our non-test modules in alphabetical order...
 
 drake_py_library(
-    name = "acrobot_io",
-    srcs = ["acrobot_io.py"],
-    deps = [
-        "//bindings/pydrake",
-    ],
-)
-
-drake_py_library(
     name = "metrics",
     srcs = ["metrics.py"],
 )
@@ -36,63 +28,14 @@ drake_py_binary(
         ":spong_sim_main_cc",
         ":test/example_stochastic_scenario.yaml",
     ],
+    tags = ["manual"],  # Manual while its dependencies move out of dev
     deps = [
         ":acrobot_io",
         ":metrics",
     ],
 )
 
-drake_py_library(
-    # This is the library form of spong_sim.py.
-    name = "spong_sim_lib_py",
-    srcs = ["spong_sim.py"],
-    deps = [
-        ":acrobot_io",
-        "//bindings/pydrake",
-    ],
-)
-
-drake_py_binary(
-    # This is the directly-runnable form of spong_sim.py.
-    name = "spong_sim_main_py",
-    srcs = ["spong_sim.py"],
-    main = "spong_sim.py",
-    deps = [
-        ":acrobot_io",
-        "//bindings/pydrake",
-    ],
-)
-
-drake_cc_binary(
-    name = "spong_sim_main_cc",
-    srcs = ["spong_sim.cc"],
-    deps = [
-        "//common:add_text_logging_gflags",
-        "//common:name_value",
-        "//common/schema:stochastic",
-        "//common/yaml:yaml_read_archive",
-        "//common/yaml:yaml_write_archive",
-        "//examples/acrobot:acrobot_plant",
-        "//examples/acrobot:spong_controller",
-        "//systems/analysis:simulator",
-        "//systems/framework:diagram_builder",
-        "//systems/primitives:signal_logger",
-        "@yaml_cpp",
-    ],
-)
-
 # === test/ ===
-
-drake_py_unittest(
-    name = "acrobot_io_test",
-    data = [
-        ":test/example_scenario.yaml",
-    ],
-    deps = [
-        ":acrobot_io",
-        "//bindings/pydrake",
-    ],
-)
 
 drake_py_unittest(
     name = "metrics_test",
@@ -105,45 +48,9 @@ sh_test(
     name = "optimizer_demo_smoke_test",
     srcs = ["test/optimizer_demo_smoke_test.sh"],
     data = [":optimizer_demo"],
-    tags = ["no_memcheck"],  # This wraps python; python is memcheck-exempt.
-)
-
-drake_py_unittest(
-    name = "spong_sim_lib_py_test",
-    deps = [
-        ":spong_sim_lib_py",
-    ],
-)
-
-drake_py_test(
-    name = "spong_sim_main_py_test",
-    srcs = ["test/spong_sim_main_test.py"],
-    allow_import_unittest = True,
-    data = [
-        ":spong_sim_main_py",
-        ":test/example_scenario.yaml",
-    ],
-    main = "test/spong_sim_main_test.py",
-    deps = [
-        ":acrobot_io",
-        "//bindings/pydrake",
-    ],
-)
-
-drake_py_test(
-    name = "spong_sim_main_cc_test",
-    srcs = ["test/spong_sim_main_test.py"],
-    allow_import_unittest = True,
-    args = ["--cc"],
-    data = [
-        ":spong_sim_main_cc",
-        ":test/example_scenario.yaml",
-        ":test/example_stochastic_scenario.yaml",
-    ],
-    main = "test/spong_sim_main_test.py",
-    deps = [
-        ":acrobot_io",
-        "//bindings/pydrake",
+    tags = [
+        "manual",  # Manual while its dependencies move out of dev
+        "no_memcheck",  # This wraps python; python is memcheck-exempt.
     ],
 )
